[MCCAS][nfc] Move DWARF helpers into separate header
Some of this functionality will be needed by the CAS dump tool as well,
therefore it must be visible.
